more pclint warnings fixed

+++ include/args.h
@@ -19,9 +19,9 @@
 class Args
 {
 protected:
-   int   m_count;       /**< number of command line arguments */
-   char  **m_values;    /**< pointer array to each argument */
-   UINT8 *m_used;       /**< bit array with one flag per argument */
+   size_t  m_count;      /**< number of command line arguments */
+   char   **m_values;    /**< pointer array to each argument */
+   UINT8  *m_used;       /**< bit array with one flag per argument */
 
    static size_t NumberOfBits(const int argc);
