+++ src/semicolons.cpp
@@ -115,7 +115,7 @@ void remove_extra_semicolons(void)
 static void check_unknown_brace_close(chunk_t *semi, chunk_t *brace_close)
 {
    LOG_FUNC_ENTRY();
-   chunk_t *pc = chunk_get_prev_type(brace_close, CT_BRACE_OPEN, brace_close->level);
+   chunk_t *pc = chunk_get_prev_type(brace_close, CT_BRACE_OPEN, (int)brace_close->level);
    pc = chunk_get_prev_ncnl(pc);
    if ((pc != NULL) &&
        (pc->type != CT_RETURN) &&
